requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argcomplete==1.10.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.8.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
coloredlogs==15.0.1
compressed-rtf==1.0.6
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.1
docx2txt==0.8
ebcdic==1.1.1
effdet==0.4.1
emoji==2.8.0
environs==9.5.0
et-xmlfile==1.1.0
extract-msg==0.28.7
faiss-cpu==1.7.4
filelock==3.13.0
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
grpcio==1.58.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
IMAPClient==2.1.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.325
langdetect==1.0.9
langsmith==0.0.53
layoutparser==0.3.4
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
minio==7.1.17
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
olefile==0.46
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
pdf2image==1.16.3
pdfminer.six==20191110
pdfplumber==0.10.3
Pillow==10.1.0
portalocker==2.8.2
protobuf==4.24.4
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pymilvus==2.3.2
pypandoc==1.12
pyparsing==3.1.1
pypdf==3.16.4
pypdfium2==4.22.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scipy==1.11.3
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.22
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
textract==1.6.5
tiktoken==0.5.1
timm==0.9.8
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
unstructured==0.10.27
unstructured-inference==0.7.10
unstructured.pytesseract==0.3.12
urllib3==2.0.7
xlrd==1.2.0
XlsxWriter==3.1.9
yarl==1.9.2
